refactor: refactor date and time to .tsx format and separated to components
