+++ src/pages/index.tsx
@@ -8,6 +8,7 @@ import { Container } from "@mui/system";
 import { createTheme } from "@mui/material";
 import { ThemeProvider } from "@emotion/react";
 import Clock from "../components/Clock";
+import Dates from "../components/Date";
 import useSWR from "swr";
 
 import "@fontsource/open-sans";
@@ -15,11 +16,6 @@ import "@fontsource/mohave";
 import "@fontsource/montserrat";
 import IEvent from "../interfaces/models/event";
 import BookCard from "../components/BookCard";
-
-const current = new Date();
-const date = `${current.getDate()}/${
-  current.getMonth() + 1
-}/${current.getFullYear()}`;
 
 const theme = createTheme({
   typography: {
@@ -52,7 +48,7 @@ const Home: NextPage = () => {
       <ThemeProvider theme={theme}>
         <Grid container justifyContent="flex-end">
           <Grid item style={{ marginTop: "2%" }}>
-            <Typography fontFamily="Open Sans">Brisbane, {date}</Typography>
+            <Dates />
             <Clock />
           </Grid>
         </Grid>
@@ -92,7 +88,11 @@ const Home: NextPage = () => {
             <Typography
               fontFamily={"Mohave"}
               fontSize={"2.2rem"}
-              style={{ color: "#784CF4", fontWeight: "bold" }}
+              style={{
+                color: "#784CF4",
+                fontWeight: "bold",
+                justifyContent: "center",
+              }}
             >
               UPCOMING EVENTS
             </Typography>
